requirements.txt
asgiref==3.4.1
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.0.1
cmake==3.20.5
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-import-export==2.5.0
dlib==19.22.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
joblib==1.0.1
MarkupPy==1.14
mathparse==0.1.2
mock==4.0.3
nltk==3.6.2
numpy==1.21.0
odfpy==1.4.1
opencv-python==4.5.2.54
openpyxl==3.0.7
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
Pint==0.17
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.7.6
regx==1.0.0
six==1.16.0
SQLAlchemy==1.2.19
sqlparse==0.4.1
tablib==3.0.0
termcolor==1.1.0
tqdm==4.61.2
xlrd==2.0.1
xlwt==1.3.0
